+++ CHANGES.rst
@@ -11,6 +11,7 @@ development
 - Honor "--use-application-id" also when printing results to STDOUT in bulk mode
 - Fix Celery task reject behavior
 - Don't use number from "--numberfile" when prefixed with "#"
+- Improve exception handling and error response for bulk acquisition mode
 
 2017-10-22 0.4.0
 ================